pyproject.toml
[project]
name = "dds-testcase"
version = "0.1.0"
description = ""
authors = [
    {name = "Vyacheslav",email = "slava20031112@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "django-admin-rangefilter (>=0.13.2,<0.14.0)",
    "django-autocomplete-light (>=3.12.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
